test(auth): extract reusable test suite
