+++ src/tests/specs/satellite/stock/auth/satellite.auth.upgrade.spec.ts
@@ -2,15 +2,11 @@ import type { SatelliteActor } from '$declarations';
 import type { Ed25519KeyIdentity } from '@dfinity/identity';
 import type { Actor, PocketIc } from '@dfinity/pic';
 import type { Principal } from '@dfinity/principal';
-import { mockClientId } from '../../../../mocks/jwt.mocks';
-import { stopCanister } from '../../../../utils/ic-management-tests.utils';
-import { makeMockGoogleOpenIdJwt } from '../../../../utils/jwt-tests.utils';
-import { assertOpenIdHttpsOutcalls } from '../../../../utils/observatory-openid-tests.utils';
-import { tick } from '../../../../utils/pic-tests.utils';
-import { setupSatelliteAuth } from '../../../../utils/satellite-auth-tests.utils';
+import { testAuthUpgrade } from '../../../../utils/auth-assertions-upgrade-tests.utils';
+import { setupSatelliteAuth, type TestSession } from '../../../../utils/auth-tests.utils';
 import { upgradeSatellite } from '../../../../utils/satellite-upgrade-tests.utils';
 
-describe('Satellite > Auth > Upgrade', () => {
+describe('Satellite > Auth', () => {
 	let pic: PocketIc;
 
 	let observatoryId: Principal;
@@ -20,16 +16,14 @@ describe('Satellite > Auth > Upgrade', () => {
 	let satelliteActor: Actor<SatelliteActor>;
 	let satelliteId: Principal;
 
-	let publicKey: Uint8Array;
-	let nonce: string;
-	let salt: Uint8Array;
+	let session: TestSession;
 
 	beforeAll(async () => {
 		const {
 			pic: p,
 			satellite: { actor: sActor, canisterId: sId },
 			observatory: { canisterId: oId },
-			session: { nonce: n, publicKey: pK, salt: sT },
+			session: s,
 			controller: c
 		} = await setupSatelliteAuth();
 
@@ -41,55 +35,20 @@ describe('Satellite > Auth > Upgrade', () => {
 
 		controller = c;
 
-		nonce = n;
-		publicKey = pK;
-		salt = sT;
+		session = s;
 	});
 
 	afterAll(async () => {
 		await pic?.tearDown();
 	});
 
-	it('should use cached certificate after upgrade', async () => {
-		await pic.advanceTime(15 * 60_000);
-		await tick(pic);
-
-		// Authenticate
-		const now = await pic.getTime();
-
-		const { jwks, jwt } = await makeMockGoogleOpenIdJwt({
-			clientId: mockClientId,
-			date: new Date(now),
-			nonce
-		});
-
-		await assertOpenIdHttpsOutcalls({ pic, jwks });
-
-		const { authenticate_user } = satelliteActor;
-
-		const result = await authenticate_user({
-			OpenId: { jwt, session_key: publicKey, salt }
-		});
-
-		expect('Ok' in result).toBeTruthy();
-
-		// Stop the Observatory. If there is a fetch it would throw an error.
-		await stopCanister({
-			canisterId: observatoryId,
-			pic,
-			sender: controller
-		});
-
-		// Upgrade Satellite
-		await upgradeSatellite({ canisterId: satelliteId, pic, controller });
-
-		await tick(pic);
-
-		// Try to re-authenticate
-		const newResult = await authenticate_user({
-			OpenId: { jwt, session_key: publicKey, salt }
-		});
-
-		expect('Ok' in newResult).toBeTruthy();
+	testAuthUpgrade({
+		pic: () => pic,
+		actor: () => satelliteActor,
+		canisterId: () => satelliteId,
+		controller: () => controller,
+		session: () => session,
+		observatoryId: () => observatoryId,
+		upgrade: () => upgradeSatellite({ canisterId: satelliteId, pic, controller })
 	});
 });
